fix support for showWarnings option in eclipse compiler
Previous implementation of turning off warnings was trying to achieve this
by enabling OPTION_SuppressWarnings, but it was wrong. Enabling this
option tell eclipse to honor @SuppressWarning annotations only, and does
not turn off all other warnings.

To do this we have to iterate thru all the options, and change those that
were set as WARNING to IGNORE. This is how the do it for example in
eclipse batch compiler.