Activate 'streamLogsOnFailures' for Maven Invoker
This way, it is easier to debug failing integration tests on CI
environments, where it is difficult or impossible to add the individual
IT build logs.

OTOH, the Maven debug statements in the test output are usually making
it quite hard to find the actual error that occurred. While helpful in
rare cases, most of the time the normal Maven build output should easily
suffice in order to identify and fix integration test problems.
Therefore, I deactivated 'debug' for Maven Invoker. This should strike a
sensible balance between inlining IT error logs into the outer Maven
build log in case of errors and not polluting it with too much
information.