Avoid NPE in AspectJCompilerTest on AspectJ 1.9.8+
Due to changed Eclipse Compiler (JDT Core) upstream of AJC, we need to
call 'buildArgParser.populateBuildConfig(..)' in order to avoid an NPE
when AjBuildConfig.getCheckedClasspaths() is called later during
compilation.

Only the unit test (which is not really a unit test) is affected by this
problem, because it compiles in-process. The 'aspectj-compiler' IT was
still passing before this change.