JavacCompiler.java had a space in the wrong place
It looks like the style rule is that annotations are written `@Foo( param="spaces in between parens and params" )`, but here, it the space was _within_ the string literal. Which would break everything, except evidently elsewhere these compilerId strings are perhaps trimmed, or, if not, the META-INF serviceloader infra trims these. Still, it was weird, and ugly.