Report "Error occurred during initialization of VM" as error
Until now, this error message was just swallowed silently.

Along the way, also report "Error occurred during initialization of boot
layer" as ERROR, because probably OTHER was never the right category to
begin with. With OTHER, Maven Compiler logs both error messages on INFO,
which I believe to be wrong. Now, Maven Compiler
logs them on ERROR with "COMPILATION ERROR" header, which fits the
behaviour that the build fails. Besides, javadoc for
CompilerMessage.Kind.ERROR says "Problem which prevents the tool's
normal completion", which also is a good description of what is actually
happening for both the boot layer and VM init errors.