Make Eclipse and AspectJ compilers run on JDK 11 again
AspectJ can compile many projects on JDK 11, because it is not always
using Java 17 classes from JDT Core. If it does not work, there will be
a runtime error, and the user can figure out what is wrong.

For ECJ, the case is more complicated, because it directly imports Java
17 classes from JDT Core. There were isolated in new class
EclipseJavaCompilerDelegate, which is accessed using Class::forName and
method handles from EclipseJavaCompiler.

I.e., Sisu Inject can scan the annotations on EclipseJavaCompiler on
JDK 11, and there is no more confusing "No such compiler 'eclipse'"
error, but rather:

Error injecting constructor, ... EcjFailureException:
Failed to run the ecj compiler: ECJ only works on Java 17+
  at org.codehaus.plexus.compiler.eclipse.EclipseJavaCompiler.<init>

This explicitly tells the user what is wrong.

A multi-release JAR solution was tested and worked well in Maven, but
was dismissed due to the terrible developer experience in IDEs.

Fixes #347.