Throw CompilerException in performCompile|getEcj, not in constructor
This shifts the focus from a supposed Maven component wiring error to
the root cause, expressed in the error message:

  Fatal error compiling: ECJ needs JRE 17+

We need to add safeguards throwing the exception in both methods
accessing class EclipseJavaCompilerDelegate, because we cannot predict
who calls which one first. Omitting one safeguard might lead to the
return of the infamous "No such compiler 'eclipse'" error.

The new static boolean field EclipseJavaCompiler.isJdkSupported also
enables us to write a simple unit test, asserting on the error message,
if the field value is false.

Relates to #347.