Code review: throw exception with cause
In order to be able to do that at all, I had to add a constructor taking
a throwable first. Now, even though a cause is propagated, at the time
of writing this Maven Compiler will just catch the
NoSuchCompilerException we throw, ignore its message and root cause and
throw a new MojoExecutionException instead. :-/

Relates to #347.