Lazy providers and better error reporting (#361)
* ITs need plugin dependency to plexus-compiler-manager

While testing #347, changes in compiler manager were not pulled into
ITs, because Maven Compiler has a dependency on it, which must be
overridden in all ITs or in projects using Plexus Compiler generally, if
they need to override the version predefined by Maven Compiler.

* Lazy providers and better error reporting

If scanning, injection or construction fails, log a comprehensive error
message on top of throwing a NoSuchCompilerException.

Fixes #347.

Co-authored-by: Alexander Kriegisch <Alexander@Kriegisch.name>

* Code review: throw exception with cause

In order to be able to do that at all, I had to add a constructor taking
a throwable first. Now, even though a cause is propagated, at the time
of writing this Maven Compiler will just catch the
NoSuchCompilerException we throw, ignore its message and root cause and
throw a new MojoExecutionException instead. :-/

Relates to #347.

* Code review: improve DefaultCompilerManager.ERROR_MESSAGE

Add more detail concerning possible user errors like misspelling the
compiler ID or missing dependencies for a compiler.

Relates to #347.

---------

Co-authored-by: Tamas Cservenak <tamas@cservenak.net>